chore: better error handling
