+++ apps/api/http/storage/upload-object.http
@@ -1,0 +1,29 @@
+### Login
+# @name login
+POST http://localhost:3000/api/v1/auth/emailpass
+Authorization: Bearer {{$dotenv TESTING_API_KEY}}
+Content-Type: application/json
+
+{
+"email": "test@example.com",
+ "password": "testing124"
+}
+
+### Get Video Presigned URL
+# @name getPresignedUrl
+POST http://localhost:3000/api/v1/storage/presigned-url/upload
+Authorization: Bearer {{$dotenv TESTING_API_KEY}}
+X-User: {{login.response.body.data.token}}
+Content-Type: application/json
+
+{
+ "filename": "sample.mp4",
+ "patientId": "1235678"
+}
+
+### Upload Video
+# @name uploadVideo
+PUT {{getPresignedUrl.response.body.data.url}}
+Content-Type: multipart/form-data
+
+< sample.mp4
