--- apps/api/http/video/get-video-presigned-url.http
@@ -1,17 +1,0 @@
-### Login
-# @name login
-POST http://localhost:3000/api/v1/auth/emailpass
-Authorization: Bearer {{$dotenv TESTING_API_KEY}}
-Content-Type: application/json
-
-{
-"email": "test@example.com",
- "password": "testing124"
-}
-
-### Get Video Presigned URL
-
-GET http://localhost:3000/api/v1/storage/videos/presigned-url/1234567/Test.mp4
-Authorization: Bearer {{$dotenv TESTING_API_KEY}}
-Content-Type: application/json
-X-User: {{login.response.body.data.token}}
