--- apps/api/http/video/put-video-presigned-url.http
@@ -1,23 +1,0 @@
-### Login
-# @name login
-POST http://localhost:3000/api/v1/auth/emailpass
-Authorization: Bearer {{$dotenv TESTING_API_KEY}}
-Content-Type: application/json
-
-{
-"email": "test@example.com",
- "password": "testing124"
-}
-
-### Get Video Presigned URL
-
-POST http://localhost:3000/api/v1/storage/videos/presigned-url
-Authorization: Bearer {{$dotenv TESTING_API_KEY}}
-Content-Type: application/json
-X-User: {{login.response.body.data.token}}
-
-{
- "date": "{{$datetime iso8601}}",
- "filename": "test.mp4",
- "patientId": "abc-123-456-789"
-}
