--- apps/api/http/video/upload-video.http
@@ -1,25 +1,0 @@
-### Login
-# @name login
-POST http://localhost:3000/api/v1/auth/emailpass
-Authorization: Bearer {{$dotenv TESTING_API_KEY}}
-Content-Type: application/json
-
-{
-"email": "test@example.com",
- "password": "testing124"
-}
-
-### Get Video Presigned URL
-# @name getPresignedUrl
-POST http://localhost:3000/api/v1/storage/videos/presigned-url
-Authorization: Bearer {{$dotenv TESTING_API_KEY}}
-Content-Type: application/json
-X-User: {{login.response.body.data.token}}
-
-{
- "filename": "{{$datetime iso8601}}.txt",
- "patientId": "abc-123-456-789"
-}
-
-### 
-PUT {{getPresignedUrl.response.body.data.url}}
