+++ apps/api/src/config/common-headers.ts
@@ -1,0 +1,3 @@
+export const COMMON_HEADERS = {
+    CONTENT_TYPE_JSON: { "Content-Type": "application/json" },
+}
