--- apps/api/src/handlers/storage/videos.ts
@@ -1,161 +1,0 @@
-import { zValidator } from "@hono/zod-validator";
-import type { Variables } from "@src/index";
-import { dbMiddleware } from "@src/middleware/db";
-import { userMiddleware } from "@src/middleware/user";
-import { validateApiKey } from "@src/middleware/validate-api-key";
-import { storageService } from "@src/services/storage";
-import {
-  presignedUrlResponseSchema,
-  presignedUrlSchema,
-} from "@src/validators/storage";
-import { Hono } from "hono";
-import { describeRoute } from "hono-openapi";
-import { resolver } from "hono-openapi/zod";
-import { stream, streamText, streamSSE } from 'hono/streaming'
-import Stream from "stream";
-import { z } from "zod";
-
-const app = new Hono<{ Variables: Variables }>().post(
-  "/presigned-url/upload",
-  describeRoute({
-    summary: "Generate a presigned URL for uploading a video",
-    description:
-      "Generate a presigned URL for uploading a video to the storage bucket",
-    tags: ["storage"],
-    requestBody: {
-      content: {
-        "application/json": {
-          schema: resolver(presignedUrlSchema),
-        },
-      },
-    },
-    responses: {
-      200: {
-        description: "Presigned URL generated successfully",
-        content: {
-          "application/json": {
-            schema: resolver(presignedUrlResponseSchema),
-          },
-        },
-      },
-    },
-  }),
-  userMiddleware,
-  dbMiddleware,
-  validateApiKey,
-  zValidator("json", presignedUrlSchema),
-  async (c) => {
-    const { filename, patientId } = c.req.valid("json");
-    const userId = c.get("userId")!;
-
-    const data = await storageService.generatePresignedVideoPutUrl(
-      filename,
-      userId,
-      patientId,
-    );
-
-    return c.json({
-      data: { ...data },
-      error: null,
-    });
-  },
-).post("/presigned-url", userMiddleware, dbMiddleware, validateApiKey, zValidator("json", presignedUrlSchema), async (c) => {
-
-  const { filename, patientId } = c.req.valid("json");
-  const userId = c.get("userId")!;
-
-  const data = await storageService.generatePresignedGetUrl(
-    filename,
-    patientId,
-    userId,
-  );
-
-  return c.json({
-    data: { ...data },
-    error: null,
-  });
-}).get(
-  "/",
-  describeRoute({
-    summary: "Get a presigned URL for uploading a video",
-    description:
-      "Get a presigned URL for uploading a video to the storage bucket",
-    tags: ["storage"],
-    parameters: [
-      {
-        patientId: {
-          description: "The ID of the patient",
-          required: true,
-          type: "string",
-        }
-      },
-      {
-        filename: {
-          description: "The name of the file to upload",
-          required: true,
-          type: "string",
-        }
-      }
-    ],
-    responses: {
-      200: {
-        description: "Presigned URL generated successfully",
-        content: {
-          "application/json": {
-            schema: resolver(presignedUrlResponseSchema),
-          },
-        },
-      },
-    },
-  }),
-  dbMiddleware,
-  zValidator("query", z.object({
-    "X-MSWA-Filename": z.string(),
-    "X-MSWA-Bucket": z.string(),
-    "X-MSWA-UserId": z.string(),
-    "X-MSWA-Method": z.string(),
-    "X-MSWA-Expires": z.string(),
-    "X-MSWA-Signature": z.string(),
-  })),
-  async (c) => {
-    const filename = c.req.valid("query")["X-MSWA-Filename"];
-    const patientId = c.req.valid("query")["X-MSWA-Bucket"];
-    const presignedUserId = c.req.valid("query")["X-MSWA-UserId"];
-    const method = c.req.valid("query")["X-MSWA-Method"];
-    const expires = c.req.valid("query")["X-MSWA-Expires"];
-    const signature = c.req.valid("query")["X-MSWA-Signature"];
-
-    await storageService.validatePresignedUrl(
-      filename,
-      patientId,
-      presignedUserId,
-      method,
-      expires,
-      signature,
-    );
-
-    const data = await storageService.getObjectStream(
-      filename,
-      patientId,
-    );
-
-    // Assuming data is a ReadableStream<Uint8Array> or similar
-    if (!data) {
-      return c.notFound();
-    }
-
-    c.header('Content-Type', 'video/mp4');
-    c.header('Accept-Ranges', 'bytes');
-
-    const nodeReadable = data.stream as Stream.Readable;
-    // Convert Node.js Readable to Web ReadableStream
-    const webReadableStream = Stream.Readable.toWeb(nodeReadable);
-
-    return stream(c, async (stream) => {
-      // Pipe the Web ReadableStream
-      await stream.pipe(webReadableStream as ReadableStream<any>);
-    })
-  },
-);
-
-export default app;
