+++ apps/api/src/repositories/refresh-token.ts
@@ -1,6 +1,7 @@
 import { refreshTokens } from "@db/schema/auth";
 import { signJWT } from "@lib/jwt";
 import type { DB } from "@middleware/db";
+import { COMMON_HEADERS } from "@src/config/common-headers";
 import { HTTP_CODES } from "@src/config/http-codes";
 import { eq, sql } from "drizzle-orm";
 import { HTTPException } from "hono/http-exception";
@@ -38,8 +39,14 @@ const createRefreshToken = async (
     return data[0].token;
   } catch (e) {
     console.error(`Failed to create refresh token: ${e}`);
-    throw new HTTPException(HTTP_CODES.NOT_IMPLEMENTED, {
-      message: "Failed to create refresh token",
+    throw new HTTPException(HTTP_CODES.INTERNAL_SERVER_ERROR, {
+      res: new Response(
+        JSON.stringify({
+          data: null,
+          error: "Failed to create refresh token",
+        }),
+        { headers: COMMON_HEADERS.CONTENT_TYPE_JSON },
+      ),
     });
   }
 };
@@ -70,8 +77,14 @@ const revokeRefreshToken = async (db: DB, token: string) => {
       .where(eq(refreshTokens.token, token));
   } catch (e) {
     console.error(`Failed to revoke refresh token: ${e}`);
-    throw new HTTPException(HTTP_CODES.NOT_IMPLEMENTED, {
-      message: "Failed to revoke refresh token",
+    throw new HTTPException(HTTP_CODES.INTERNAL_SERVER_ERROR, {
+      res: new Response(
+        JSON.stringify({
+          data: null,
+          error: "Failed to revoke refresh token",
+        }),
+        { headers: COMMON_HEADERS.CONTENT_TYPE_JSON },
+      ),
     });
   }
 };
